added getDuration util in health-dss module (#2708)
