+++ health/micro-ui/web/micro-ui-internals/packages/modules/health-dss/src/components/CustomPieChart.js
@@ -5,6 +5,7 @@ import FilterContext from "./FilterContext";
 import NoData from "./NoData";
 import { getTitleHeading } from "../utils/locale";
 import { Loader, Chip } from "@egovernments/digit-ui-components";
+import { getDuration } from "../utils/getDuration";
 const COLORS = ["#048BD0", "#FBC02D", "#8E29BF", "#EA8A3B", "#0BABDE", "#6E8459", "#D4351C", "#0CF7E4", "#F80BF4", "#22F80B"];
 const mobileView = innerWidth <= 640;
 
@@ -12,7 +13,7 @@ const getInitialRange = () => {
   const data = Digit.SessionStorage.get("DSS_FILTERS");
   const startDate = data?.range?.startDate ? new Date(data?.range?.startDate) : Digit.Utils.dss.getDefaultFinacialYear().startDate;
   const endDate = data?.range?.endDate ? new Date(data?.range?.endDate) : Digit.Utils.dss.getDefaultFinacialYear().endDate;
-  const interval = Digit.Utils.dss.getDuration(startDate, endDate);
+  const interval = getDuration(startDate, endDate);
   return { startDate, endDate, interval };
 };
