+++ health/micro-ui/web/micro-ui-internals/packages/modules/health-dss/src/components/MapChart.js
@@ -6,6 +6,7 @@ import { get } from "lodash";
 import { Loader } from "@egovernments/digit-ui-components";
 import { ResponsiveContainer } from "recharts";
 import { format } from "date-fns";
+import { getDuration } from "../utils/getDuration";
 
 const PROJECTION_CONFIG = { scale: 320, center: [85.9629, 22.5937] };
 
@@ -21,7 +22,7 @@ const getInitialRange = () => {
   const startDate = data?.range?.startDate ? new Date(data?.range?.startDate) : Digit.Utils.dss.getDefaultFinacialYear().startDate;
   const endDate = data?.range?.endDate ? new Date(data?.range?.endDate) : Digit.Utils.dss.getDefaultFinacialYear().endDate;
   const title = `${format(startDate, "MMM d, yyyy")} - ${format(endDate, "MMM d, yyyy")}`;
-  const interval = Digit.Utils.dss.getDuration(startDate, endDate);
+  const interval = getDuration(startDate, endDate);
   const denomination = data?.denomination || "Lac";
   const tenantId = data?.filters?.tenantId || [];
   return { startDate, endDate, title, interval, denomination, tenantId };
