+++ health/micro-ui/web/micro-ui-internals/packages/modules/health-dss/src/components/StackedTable.js
@@ -5,6 +5,7 @@ import { getTitleHeading } from "../utils/locale";
 import FilterContext from "./FilterContext";
 import { Button } from "@egovernments/digit-ui-components";
 import { useHistory } from "react-router-dom";
+import { getDuration } from "../utils/getDuration";
 
 export default function StackedTable({ chartId, visualizer, initialRange, isNational, routeTo, redirectUrl }) {
   const { t } = useTranslation();
@@ -25,7 +26,7 @@ export default function StackedTable({ chartId, visualizer, initialRange, isNati
     const data = Digit.SessionStorage.get("DSS_FILTERS");
     const startDate = data?.range?.startDate ? new Date(data?.range?.startDate) : Digit.Utils.dss.getDefaultFinacialYear().startDate;
     const endDate = data?.range?.endDate ? new Date(data?.range?.endDate) : Digit.Utils.dss.getDefaultFinacialYear().endDate;
-    const interval = Digit.Utils.dss.getDuration(startDate, endDate);
+    const interval = getDuration(startDate, endDate);
     return { startDate, endDate, interval };
   };
