+++ health/micro-ui/web/micro-ui-internals/packages/modules/health-dss/src/utils/getDuration.js
@@ -1,0 +1,15 @@
+export const getDuration = (startDate, endDate) => {
+  let noOfDays = (new Date(endDate).getTime() - new Date(startDate).getTime()) / (1000 * 3600 * 24);
+  if (noOfDays >= 90) {
+    console.log("inside month")
+    return "month";
+  }
+  if (noOfDays < 90 && noOfDays >= 14) {
+    console.log("inside week")
+    return "week";
+  }
+  if (noOfDays < 14) {
+    console.log("inside day")
+    return "day";
+  }
+};
